core/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
josepy==1.14.0
mozilla-django-oidc==4.0.1
packaging==24.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
